refactor: Integrate domain and entity models
